added alter migrations

+++ server/db/migrations/triviaOne.js
@@ -2,12 +2,12 @@ exports.up = (knex, Promise) => {
   return knex.schema.createTable('trivia_one', (table) => {
     table.increments('id').primary()
     table.integer('user_id')
-    table.decimal('brier_score_one')
-    table.decimal('brier_score_two')
-    table.decimal('brier_score_three')
-    table.decimal('brier_score_four')
-    table.decimal('brier_score_five')
-    table.decimal('brier_score_total')
+    table.integer('brier_score_one')
+    table.integer('brier_score_two')
+    table.integer('brier_score_three')
+    table.integer('brier_score_four')
+    table.integer('brier_score_five')
+    table.integer('brier_score_total')
   })
 }
